Move make_mock_urls to osprey/testing and import in conftest
